Add method to disable ABC feature of B version sensor
This is just a suggestion of a minimum to do - you will probably want to re-write it into the init or where it makes sense to catch the point of activating.  My assumption is that you never need to enable (probably wrong).  I think this may be programmed into the firmware and re-enabled with any reset, self initiated or external (power cycle, etc.) I'll bet lots of folks think they have a defective sensor after some time operating, assuming they ever question the data.