Update on_off_kasa_plugs.py
The "aio_rpc_server" is passed four arguments in line 176/177, but the "aio_rpc_server()" function only uses three arguments.
This results in the rpc server to not start at all and leaving no log error whatsoever. The output setup will fail and a Kasa plugs cannot be controlled.
By adding a fourth argument "channel_count" the rpc server is started correctly and the output module for Kasa plugs is working fine.
It should be noted that new argument is never used but turning on/off the plugs works fine.