Solution to Issue #1234
Change argument passed to force_calibration method from float to int. The library is expecting an int as can be seen here:
https://github.com/adafruit/Adafruit_CircuitPython_SCD4X/blob/a9ae07d8c30c45e7a36d5308bdaaa2887c3c7e31/adafruit_scd4x.py#L161

This fix restarts the device so that it automatically starts working without needing to deactivate and activate after force calibrating CO2.

NOTE: The adafruit library calls `sensor.stop_periodic_measurement()` at the begining of the function but never calls `sensor.start_periodic_measurement()` after it submits the calibration.